move grpc client to a separate package
