+++ aggsender/flows/flow_aggchain_prover.go
@@ -12,14 +12,14 @@ import (
 	"github.com/agglayer/aggkit/aggsender/grpc"
 	"github.com/agglayer/aggkit/aggsender/types"
 	"github.com/agglayer/aggkit/bridgesync"
-	aggkitcommon "github.com/agglayer/aggkit/common"
+	aggkitgrpc "github.com/agglayer/aggkit/grpc"
 	treetypes "github.com/agglayer/aggkit/tree/types"
 	aggkittypes "github.com/agglayer/aggkit/types"
 	"github.com/ethereum/go-ethereum/common"
 	"google.golang.org/grpc/codes"
 )
 
-var errNoProofBuiltYet = &aggkitcommon.GRPCError{
+var errNoProofBuiltYet = &aggkitgrpc.GRPCError{
 	Code:    codes.Unavailable,
 	Message: "Proposer service has not built any proof yet",
 }
